faz query para relations
