+++ constants.js
@@ -7,7 +7,10 @@ const OVERPASS_SERVERS = [
 ];
 const DEFAULT_BORDER_WIDTH = 3;
 
+const RELATION_IDS = [15980035, 15980034, 15997277, 15997277, 15999831, 15999831, 16000488, 16005002, 16005002, 16000461, 16000461, 16000461, 16000034, 16000483, 16000484, 16000484, 16002184, 16002184, 16004643, 15997436, 15997436, 15997435, 16001885, 16001884, 16001884, 16001874, 16002174, 15997447, 15997448, 15997449, 15997450, 15997451, 15997451, 16002171, 16002114, 16002114, 16002092, 16002092, 15997674, 15997675, 15997690, 15997691, 15997691, 15997694, 15997695, 15997713, 15997714, 15997715, 15997715, 15997464, 15997464, 15997465, 15997465, 15997469, 15997469, 15997469, 15997716, 15997717, 16002036, , , 16002035, 16000464, 16004411, 15997709, 15997704, 15997697, 15997276, 16000463, 15997458, 15997688, 15997692, 15997696, 15997466, 15997705, 15997706, 16006232, 15997711, 15997712, 16000032, 16002093, 16001889, 16006327, 16001883, 16002170, 16000489, 16006336, 16000466, 16000467, 15997459, 16002039, 16001872, 15997686, 15997685, 15997676, 15997681, 15997682, 15997460, 15997684, 16004565, 15999872, 15980073, 16006053, 15980070, 15980072, 16006342, 16001946, 16004999, 15997457, 15997456, 16006341, 16001873, 16001878, 15997275, 16001941, 16001939, 16001940, 16001938, 15980004, 15979999, 15980003, 15980001, 15980002, 16002257, 16005001, 16004640, 16004641, 16004642, 15997445, 16001880, 16002213, 16002215, 16002243, 15997677, 16001871, 16001870, 16004428, 16002119, 15997672, 16004645, 15997701, 15997689, 15997702, 15997703, 15980071, 15980074, 16006092, 16006025, 16001882, 16005000, 15997666, 16000468, 16000469, 15997670, 16000465, 15997667, 16002183, 16002168, 15997462, 16004524, 15997683, 16001879, 16002185, 15997444, 16001868, 16000042, 16004432, 16001881, 15997698, 15997679, 15997446, 16004574, 16004381, 16001869, 15997680, 16001875, 16004403, 16004564, 15997700, 15980036, 16001876, 16006352, 16000041, 16000040, 16000460, 16004405, 15980000, 15980005, 15997461, 15997463, 15997467, 15997455, 16004644, 15997439, 15997443, 15997442, 15997668, 16000033, 15997708, 15997707, 15999830, 16000038, 16000037, 16002259, 16004427, 16004404, 16004412, 16002116, 16004523, 10645820, 16001952, 16001953, 16001955, 16004605, 16001867, 15997687, 15997693, 16002091, 15997440, 16002038, 16002037, 16004563, 16005003, 15997454, 16002115, 15997671, 16004429, 16004410, 16004430, 16006132, 16002175, 16000039, 16000036, 16000035, 16001863, 16001864, 16001877, 16002156, 16001865, 16002258, 16004522, 16006355, 16002214, 16002034, 16002172, 16002173, 15997441, 16002096, 15997678, 15997437, 15997669, 16001992, 16002182, 16002094, 16001888, 16006390]
+
 module.exports = {
     OVERPASS_SERVERS,
-    DEFAULT_BORDER_WIDTH
+    DEFAULT_BORDER_WIDTH,
+    RELATION_IDS
 };
